Time: 18 ms (87.45%), Space: 20.7 MB (64.75%) - LeetHub

+++ 1913-maximum-product-difference-between-two-pairs/1913-maximum-product-difference-between-two-pairs.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int maxProductDifference(vector<int>& nums) {
         sort(nums.begin(),nums.end());
-        int res=(nums[nums.size()-1]*nums[nums.size()-2])-(nums[0]*nums[1]);
+       long res=(nums[nums.size()-1]*nums[nums.size()-2])-(nums[0]*nums[1]);
         return res;
         
     }
